training/requirements.txt
# TrOCR Fine-tuning Requirements
# Install with: pip install -r requirements.txt

# Core ML
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Training utilities
evaluate>=0.4.0
jiwer>=3.0.0  # For WER/CER metrics
tensorboard>=2.14.0
wandb>=0.15.0  # Optional: experiment tracking

# Data handling
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0

# Google Drive download helper
gdown>=4.7.0

# ONNX export (optional)
onnx>=1.14.0
onnxruntime>=1.16.0
optimum>=1.13.0
